requirements.txt
flask
requests
google-api-python-client

